Rope + Transitions Implemented [Basic]

+++ Project2D/Assets/Scripts/Chunk.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEditor;
 
 [CreateAssetMenu(fileName = "New Chunk", menuName = "Chunk")]
 public class Chunk : ScriptableObject
 {
-
-    public int lenght;
+    public float lenght;
 
     public GameObject prefab;
-
+    
 }
